Removal of unnecessary user insert function
We had previously expected users to log in through Tempoture and not Spotify, so we created a general function that would handle that but we no longer need it as users will only log in through Spotify. The user insert function that was left will need to be updated soon.